@@ -4,8 +4,8 @@
   "description": "React boilerplate template",
   "main": "src/index.ts",
   "scripts": {
+    "postinstall": "npm run css",
     "start": "env NODE_ENV=development webpack serve --mode development --config webpack.dev.js",
-    "docker": "env NODE_ENV=local webpack-dev-server --host 0.0.0.0 --port 3000 --mode development --config webpack.config.js",
     "dev": "env NODE_ENV=development webpack --mode development --config webpack.dev.js",
     "build": "env NODE_ENV=production webpack --mode production --config webpack.prod.js",
     "analyze": "env NODE_ENV=analyze webpack --mode development --config webpack.dev.js",
@@ -20,8 +20,8 @@
     "@babel/preset-react": "^7.12.13",
     "@babel/preset-typescript": "^7.15.0",
     "@fullhuman/postcss-purgecss": "^4.0.2",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.1",
+    "@types/react": "^18.0.33",
+    "@types/react-dom": "^18.0.11",
     "babel-loader": "^8.2.2",
     "css-loader": "^5.1.1",
     "css-minimizer-webpack-plugin": "^1.2.0",
@@ -35,17 +35,16 @@
     "terser-webpack-plugin": "^5.1.1",
     "webpack": "^5.24.3",
     "webpack-bundle-analyzer": "^4.4.0",
-    "webpack-cli": "^4.5.0",
+    "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^3.11.2",
-    "webpack-manifest-plugin": "^4.0.2",
     "webpack-merge": "^5.7.3",
     "webpackbar": "^5.0.0-3"
   },
   "dependencies": {
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.10",
     "postcss-loader": "^5.1.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-hot-loader": "^4.13.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "tailwindcss": "^2.2.7",
     "ts-loader": "^8.0.17"
   }
