@@ -1,5 +1,6 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import App from "./App";
 
-ReactDOM.render(<App />, document.getElementById("root"));
+// @ts-ignore: root element is there, trust me...
+createRoot(document.getElementById("root")).render(<App />);
