@@ -4,6 +4,7 @@ const common = require("./webpack.common.js");
 const HtmlWebpackPlugin = require("html-webpack-plugin");
 const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 const WatchMissingNodeModulesPlugin = require("react-dev-utils/WatchMissingNodeModulesPlugin");
+const ReactRefreshWebpackPlugin = require('@pmmmwh/react-refresh-webpack-plugin');
 
 module.exports = merge(common, {
   mode: "development",
@@ -33,8 +34,8 @@ module.exports = merge(common, {
       favicon: path.join(__dirname, "public/favicon.ico"),
       url: "https://localhost:3001",
       custom: {
-        react_lib: "https://unpkg.com/react@17/umd/react.development.js",
-        react_dom_lib: "https://unpkg.com/react-dom@17/umd/react-dom.development.js",
+        react_lib: "https://unpkg.com/react@18/umd/react.development.js",
+        react_dom_lib: "https://unpkg.com/react-dom@18/umd/react-dom.development.js",
       },
       minify: {
         removeComments: true,
@@ -55,6 +56,6 @@ module.exports = merge(common, {
       ignoreOrder: true,
     }),
     new WatchMissingNodeModulesPlugin(path.resolve("node_modules")),
-    // new ManifestPlugin(),
+    new ReactRefreshWebpackPlugin(),
   ],
 });
